split dependencies & resolvers into separate modules, reduce imports needed to call get_static_dependencies in default cases
