+++ cumulusci/core/dependencies/dependencies.py
@@ -1,8 +1,7 @@
 import abc
 import itertools
 import logging
-from enum import Enum
-from typing import Iterable, List, Optional, Tuple
+from typing import List, Optional
 
 import pydantic
 from github3.exceptions import NotFoundError
@@ -18,15 +17,8 @@ from cumulusci.core.dependencies.github import (
 )
 from cumulusci.core.dependencies.utils import TaskContext
 from cumulusci.core.exceptions import (
-    CumulusCIException,
     DependencyParseError,
     DependencyResolutionError,
-)
-from cumulusci.core.github import (
-    find_latest_release,
-    find_repo_2gp_context,
-    find_repo_feature_prefix,
-    get_version_id_from_commit,
 )
 from cumulusci.salesforce_api.metadata import ApiDeploy
 from cumulusci.salesforce_api.package_install import (
@@ -37,13 +29,7 @@ from cumulusci.salesforce_api.package_install import (
 )
 from cumulusci.salesforce_api.package_zip import MetadataPackageZipBuilder
 from cumulusci.utils import download_extract_github_from_repo, download_extract_zip
-from cumulusci.utils.git import (
-    construct_release_branch_name,
-    get_feature_branch_name,
-    get_release_identifier,
-    is_release_branch_or_child,
-    split_repo_url,
-)
+from cumulusci.utils.git import split_repo_url
 from cumulusci.utils.yaml.model_parser import CCIModel
 
 logger = logging.getLogger(__name__)
@@ -56,18 +42,6 @@ class HashableBaseModel(CCIModel):
     # See https://github.com/samuelcolvin/pydantic/issues/1303
     def __hash__(self):
         return hash((type(self),) + tuple(self.json()))
-
-
-class DependencyResolutionStrategy(str, Enum):
-    """Enum that defines a strategy for resolving a dynamic dependency into a static dependency."""
-
-    STATIC_TAG_REFERENCE = "tag"
-    COMMIT_STATUS_EXACT_BRANCH = "commit_status_exact_branch"
-    COMMIT_STATUS_RELEASE_BRANCH = "commit_status_release_branch"
-    COMMIT_STATUS_PREVIOUS_RELEASE_BRANCH = "commit_status_previous_release_branch"
-    BETA_RELEASE_TAG = "latest_beta"
-    RELEASE_TAG = "latest_release"
-    UNMANAGED_HEAD = "unmanaged"
 
 
 class Dependency(HashableBaseModel, abc.ABC):
@@ -99,12 +73,6 @@ class Dependency(HashableBaseModel, abc.ABC):
     def flatten(self, context: BaseProjectConfig) -> List["Dependency"]:
         """Get a list including this dependency as well as its transitive dependencies."""
         return [self]
-
-    def resolve(
-        self, context: BaseProjectConfig, strategies: List[DependencyResolutionStrategy]
-    ):
-        """Resolve a dependency that is not pinned to a specific version into one that is."""
-        pass
 
     def __str__(self):
         return self.description
@@ -140,54 +108,12 @@ class DynamicDependency(Dependency, abc.ABC):
     def is_flattened(self):
         return False
 
-    def resolve(
-        self, context: BaseProjectConfig, strategies: List[DependencyResolutionStrategy]
-    ):
-        """Try to resolve this dependency using the specified strategies.
+    def resolve(self, context, strategies):
+        """Resolve a DynamicDependency that is not pinned to a specific version into one that is."""
+        # avoid import cycle
+        from .resolvers import resolve_dependency
 
-        If successful, sets `self.ref` and optionally `self.managed_dependency`
-        (if a package release is found).
-
-        Otherwise raises DependencyResolutionError.
-        """
-
-        if self.is_resolved:
-            return
-
-        for s in strategies:
-            resolver = get_resolver(s, self)
-
-            if resolver and resolver.can_resolve(self, context):
-                try:
-                    self.ref, self.managed_dependency = resolver.resolve(self, context)
-                    if self.ref:
-                        break
-                except DependencyResolutionError:
-                    context.logger.info(
-                        f"Resolution strategy {s} failed for dependency {self}."
-                    )
-
-        if not self.ref:
-            raise DependencyResolutionError(f"Unable to resolve dependency {self}")
-
-
-class Resolver(abc.ABC):
-    """Abstract base class for dependency resolution strategies."""
-
-    name = "Resolver"
-
-    @abc.abstractmethod
-    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
-        pass
-
-    @abc.abstractmethod
-    def resolve(
-        self, dep: DynamicDependency, context: BaseProjectConfig
-    ) -> Tuple[Optional[str], Optional[StaticDependency]]:
-        pass
-
-    def __str__(self):
-        return self.name
+        resolve_dependency(self, context, strategies)
 
 
 class BaseGitHubDependency(DynamicDependency, abc.ABC):
@@ -687,387 +613,3 @@ def parse_dependency(dep_dict: dict) -> Optional[Dependency]:
                 return dep
         except pydantic.ValidationError:
             pass
-
-
-## Resolvers
-
-
-class GitHubTagResolver(Resolver):
-    """Resolver that identifies a ref by a specific GitHub tag."""
-
-    name = "GitHub Tag Resolver"
-
-    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
-        return isinstance(dep, BaseGitHubDependency) and dep.tag is not None
-
-    def resolve(
-        self, dep: BaseGitHubDependency, context: BaseProjectConfig
-    ) -> Tuple[Optional[str], Optional[StaticDependency]]:
-        try:
-            # Find the github release corresponding to this tag.
-            repo = get_repo(dep.github, context)
-            release = repo.release_from_tag(dep.tag)
-            ref = repo.tag(repo.ref(f"tags/{release.tag_name}").object.sha).object.sha
-            package_config = get_remote_project_config(repo, ref)
-            package_name, namespace = get_package_data(package_config)
-
-            if not dep.is_unmanaged and not namespace:
-                raise DependencyResolutionError(
-                    f"The tag {dep.tag} in {dep.github} does not identify a managed release"
-                )
-
-            if not dep.is_unmanaged:
-                return (
-                    ref,
-                    PackageNamespaceVersionDependency(
-                        namespace=namespace,
-                        version=release.name,
-                        package_name=package_name,
-                    ),
-                )
-            else:
-                return ref, None
-        except NotFoundError:
-            raise DependencyResolutionError(f"No release found for tag {dep.tag}")
-
-
-class GitHubReleaseTagResolver(Resolver):
-    """Resolver that identifies a ref by finding the latest GitHub release."""
-
-    name = "GitHub Release Resolver"
-    include_beta = False
-
-    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
-        return isinstance(dep, BaseGitHubDependency)
-
-    def resolve(
-        self, dep: BaseGitHubDependency, context: BaseProjectConfig
-    ) -> Tuple[Optional[str], Optional[StaticDependency]]:
-        repo = get_repo(dep.github, context)
-        release = find_latest_release(repo, include_beta=self.include_beta)
-        if release:
-            ref = repo.tag(repo.ref(f"tags/{release.tag_name}").object.sha).object.sha
-            package_config = get_remote_project_config(repo, ref)
-            package_name, namespace = get_package_data(package_config)
-
-            return (
-                ref,
-                PackageNamespaceVersionDependency(
-                    namespace=namespace, version=release.name, package_name=package_name
-                ),
-            )
-
-        return (None, None)
-
-
-class GitHubBetaReleaseTagResolver(GitHubReleaseTagResolver):
-    """Resolver that identifies a ref by finding the latest GitHub release, including betas."""
-
-    name = "GitHub Release Resolver (Betas)"
-    include_beta = True
-
-
-class GitHubUnmanagedHeadResolver(Resolver):
-    """Resolver that identifies a ref by finding the latest commit on the main branch."""
-
-    name = "GitHub Unmanaged Resolver"
-
-    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
-        return isinstance(dep, BaseGitHubDependency)
-
-    def resolve(
-        self, dep: BaseGitHubDependency, context: BaseProjectConfig
-    ) -> Tuple[Optional[str], Optional[StaticDependency]]:
-        repo = get_repo(dep.github, context)
-        return (repo.branch(repo.default_branch).commit.sha, None)
-
-
-class GitHubReleaseBranchResolver(Resolver, abc.ABC):
-    """Abstract base class for resolvers that use commit statuses on release branches to find refs."""
-
-    def can_resolve(self, dep: DynamicDependency, context: BaseProjectConfig) -> bool:
-        return self.is_valid_repo_context(context) and isinstance(
-            dep, BaseGitHubDependency
-        )
-
-    def get_release_id(self, context: BaseProjectConfig) -> int:
-        if not context.repo_branch or not context.project__git__prefix_feature:
-            raise DependencyResolutionError(
-                "Cannot get current branch or feature branch prefix"
-            )
-        release_id = get_release_identifier(
-            context.repo_branch, context.project__git__prefix_feature
-        )
-        if not release_id:
-            raise DependencyResolutionError("Cannot get current release identifier")
-
-        return int(release_id)
-
-    def is_valid_repo_context(self, context) -> bool:
-        return (
-            context.repo_branch
-            and context.project__git__prefix_feature
-            and is_release_branch_or_child(
-                context.repo_branch, context.project__git__prefix_feature
-            )
-        )
-
-    def locate_commit_status_package_id(self, remote_repo, release_branch, context_2gp):
-        version_id = None
-        count = 0
-        commit = release_branch.commit
-        while version_id is None and count < 5:
-            version_id = get_version_id_from_commit(
-                remote_repo, commit.sha, context_2gp
-            )
-            if version_id:
-                break
-            count += 1
-            if commit.parents:
-                commit = remote_repo.commit(commit.parents[0]["sha"])
-            else:
-                commit = None
-                break
-
-        return version_id, commit
-
-
-class GitHubReleaseBranchCommitStatusResolver(GitHubReleaseBranchResolver):
-    """Resolver that identifies a ref by finding a beta package version
-    in a commit status on a `feature/NNN` release branch."""
-
-    name = "GitHub Release Branch Commit Status Resolver"
-    branch_depth = 1
-
-    def resolve(
-        self, dep: BaseGitHubDependency, context: BaseProjectConfig
-    ) -> Tuple[Optional[str], Optional[StaticDependency]]:
-
-        release_id = self.get_release_id(context)
-        repo = context.get_repo_from_url(dep.github)
-        if not repo:
-            raise DependencyResolutionError(
-                f"Unable to access GitHub repository for {dep.github}"
-            )
-
-        try:
-            remote_branch_prefix = find_repo_feature_prefix(repo)
-            remote_2gp_context = find_repo_2gp_context(repo)
-        except Exception:
-            context.logger.info(
-                f"Could not find feature branch prefix or commit-status context for {repo.clone_url}. Unable to resolve packages."
-            )
-            return (None, None)
-
-        # We will check at least the release branch corresponding to our release id.
-        # We may be configured to check backwards on release branches.
-        release_branch = None
-        for i in range(0, self.branch_depth):
-            try:
-                remote_matching_branch = construct_release_branch_name(
-                    remote_branch_prefix, str(release_id - i)
-                )
-
-                release_branch = repo.branch(remote_matching_branch)
-            except NotFoundError:
-                pass
-
-        if release_branch:
-            version_id, commit = self.locate_commit_status_package_id(
-                repo,
-                release_branch,
-                remote_2gp_context,
-            )
-
-            if version_id:
-                context.logger.info(
-                    f"Located package version {version_id} for release {release_id} on {repo.clone_url} at commit {release_branch.commit.sha}"
-                )
-                package_config = get_remote_project_config(repo, commit.sha)
-                package_name, _ = get_package_data(package_config)
-
-                return commit.sha, PackageVersionIdDependency(
-                    version_id=version_id, package_name=package_name
-                )
-
-        context.logger.warn(
-            f"No package version located for release {release_id} on {repo.clone_url}."
-        )
-        return (None, None)
-
-
-class GitHubPreviousReleaseBranchCommitStatusResolver(
-    GitHubReleaseBranchCommitStatusResolver
-):
-    """Resolver that identifies a ref by finding a beta package version
-    in a commit status on a `feature/NNN` release branch that is earlier
-    than the matching local release branch."""
-
-    name = "GitHub Previous Release Branch Commit Status Resolver"
-    branch_depth = 3
-
-
-class GitHubReleaseBranchExactMatchCommitStatusResolver(GitHubReleaseBranchResolver):
-    """Resolver that identifies a ref by finding a beta package version
-    in a commit status on a branch whose name matches the local branch."""
-
-    name = "GitHub Exact-Match Commit Status Resolver"
-
-    def resolve(
-        self, dep: BaseGitHubDependency, context: BaseProjectConfig
-    ) -> Tuple[Optional[str], Optional[StaticDependency]]:
-        release_id = self.get_release_id(context)
-
-        repo = context.get_repo_from_url(dep.github)
-        if not repo:
-            raise DependencyResolutionError(
-                f"Unable to access GitHub repository for {dep.github}"
-            )
-
-        try:
-            remote_branch_prefix = find_repo_feature_prefix(repo)
-            remote_2gp_context = find_repo_2gp_context(repo)
-        except Exception:
-            context.logger.info(
-                f"Could not find feature branch prefix or commit-status context for {repo.clone_url}. Unable to resolve package."
-            )
-            return (None, None)
-
-        # Attempt exact match
-        try:
-            branch = get_feature_branch_name(
-                context.repo_branch, context.project__git__prefix_feature
-            )
-            release_branch = repo.branch(f"{remote_branch_prefix}{branch}")
-        except Exception:
-            context.logger.info(f"Exact-match branch not found for {repo.clone_url}.")
-            return (None, None)
-
-        version_id, commit = self.locate_commit_status_package_id(
-            repo,
-            release_branch,
-            remote_2gp_context,
-        )
-
-        if version_id:
-            context.logger.info(
-                f"Located package version {version_id} for release {release_id} on {repo.clone_url} at commit {release_branch.commit.sha}"
-            )
-
-            package_config = get_remote_project_config(repo, commit.sha)
-            package_name, _ = get_package_data(package_config)
-
-            return commit.sha, PackageVersionIdDependency(
-                version_id=version_id, package_name=package_name
-            )
-
-        context.logger.warn(
-            f"No package version located for release {release_id} on {repo.clone_url}."
-        )
-        return (None, None)
-
-
-RESOLVER_CLASSES = {
-    DependencyResolutionStrategy.STATIC_TAG_REFERENCE: GitHubTagResolver,
-    DependencyResolutionStrategy.COMMIT_STATUS_EXACT_BRANCH: GitHubReleaseBranchExactMatchCommitStatusResolver,
-    DependencyResolutionStrategy.COMMIT_STATUS_RELEASE_BRANCH: GitHubReleaseBranchCommitStatusResolver,
-    DependencyResolutionStrategy.COMMIT_STATUS_PREVIOUS_RELEASE_BRANCH: GitHubPreviousReleaseBranchCommitStatusResolver,
-    DependencyResolutionStrategy.BETA_RELEASE_TAG: GitHubBetaReleaseTagResolver,
-    DependencyResolutionStrategy.RELEASE_TAG: GitHubReleaseTagResolver,
-    DependencyResolutionStrategy.UNMANAGED_HEAD: GitHubUnmanagedHeadResolver,
-}
-
-
-## External API
-
-
-def get_resolver(
-    strategy: DependencyResolutionStrategy, dependency: DynamicDependency
-) -> Optional[Resolver]:
-    """Return an instance of a resolver class capable of applying the specified
-    resolution strategy to the dependency."""
-    # This will be fleshed out when further types of DynamicDependency are added.
-
-    return RESOLVER_CLASSES[strategy]()
-
-
-def get_resolver_stack(
-    context: BaseProjectConfig, name: str
-) -> List[DependencyResolutionStrategy]:
-    """Return a sequence of resolution strategies identified by the given `name`,
-    which can be either a named strategy from project__dependency_resolutions__resolution_strategies
-    or an alias like `production`."""
-    resolutions = context.project__dependency_resolutions
-    stacks = context.project__dependency_resolutions__resolution_strategies
-
-    if name in resolutions and name != "resolution_strategies":
-        name = resolutions[name]
-
-    if stacks and name in stacks:
-        return [DependencyResolutionStrategy(n) for n in stacks[name]]
-
-    raise CumulusCIException(f"Resolver stack {name} was not found.")
-
-
-def get_static_dependencies(
-    dependencies: List[Dependency],
-    strategies: List[DependencyResolutionStrategy],
-    context: BaseProjectConfig,
-    ignore_deps: List[dict] = None,
-):
-    """Resolves the project__dependencies section of cumulusci.yml
-    to convert dynamic GitHub dependencies into static dependencies
-    by inspecting the referenced repositories.
-
-    Keyword arguments:
-    :param dependencies: a list of dependencies to resolve
-    :param ignore_deps: if provided, ignore the specified dependencies wherever found.
-    """
-
-    while any(not d.is_flattened or not d.is_resolved for d in dependencies):
-        for d in dependencies:
-            if not d.is_resolved:
-                d.resolve(context, strategies)
-
-        def unique(it: Iterable):
-            seen = set()
-
-            for each in it:
-                if each not in seen:
-                    seen.add(each)
-                    yield each
-
-        dependencies = list(
-            unique(
-                itertools.chain(
-                    *list(
-                        d.flatten(context)
-                        for d in dependencies
-                        if not _should_ignore_dependency(d, ignore_deps or [])
-                    )
-                ),
-            )
-        )
-
-    # Make sure, if we had no flattening or resolving to do, that we apply the ignore list.
-    return [
-        d for d in dependencies if not _should_ignore_dependency(d, ignore_deps or [])
-    ]
-
-
-def _should_ignore_dependency(dependency: Dependency, ignore_deps: List[dict]):
-    if not ignore_deps:
-        return False
-
-    ignore_github = [d["github"] for d in ignore_deps if "github" in d]
-    ignore_namespace = [d["namespace"] for d in ignore_deps if "namespace" in d]
-
-    if (
-        isinstance(dependency, PackageNamespaceVersionDependency)
-        and dependency.namespace
-    ):
-        return dependency.namespace in ignore_namespace
-    if isinstance(dependency, BaseGitHubDependency):
-        return dependency.github in ignore_github
-
-    return False
