+++ cumulusci/core/dependencies/tests/test_dependencies.py
@@ -1,341 +1,29 @@
 from distutils.version import StrictVersion
-import functools
-from cumulusci.core.config.OrgConfig import OrgConfig, VersionInfo
 from typing import List, Optional, Tuple
 from unittest import mock
 
-from github3.exceptions import NotFoundError
 import pytest
 from pydantic import ValidationError
 
-from cumulusci.core.config import UniversalConfig
+from cumulusci.core.config.OrgConfig import OrgConfig, VersionInfo
 from cumulusci.core.config.project_config import BaseProjectConfig
 from cumulusci.core.dependencies.dependencies import (
-    DependencyResolutionStrategy,
     DynamicDependency,
-    GitHubBetaReleaseTagResolver,
     GitHubDynamicDependency,
     GitHubDynamicSubfolderDependency,
-    GitHubReleaseBranchCommitStatusResolver,
-    GitHubReleaseBranchExactMatchCommitStatusResolver,
-    GitHubReleaseBranchResolver,
-    GitHubReleaseTagResolver,
-    GitHubTagResolver,
-    GitHubUnmanagedHeadResolver,
     PackageNamespaceVersionDependency,
     PackageVersionIdDependency,
-    Resolver,
     StaticDependency,
     UnmanagedGitHubRefDependency,
     UnmanagedZipURLDependency,
-    get_resolver,
-    get_resolver_stack,
-    get_static_dependencies,
     parse_dependency,
 )
-from cumulusci.core.exceptions import CumulusCIException, DependencyResolutionError
-from cumulusci.core.tests.test_config import (
-    DummyContents,
-    DummyGithub,
-    DummyRelease,
-    DummyRepository,
-)
+from cumulusci.core.dependencies.resolvers import DependencyResolutionStrategy, Resolver
+from cumulusci.core.exceptions import DependencyResolutionError
 from cumulusci.salesforce_api.package_install import (
     DEFAULT_PACKAGE_RETRY_OPTIONS,
     PackageInstallOptions,
 )
-from cumulusci.utils.git import split_repo_url
-
-
-@pytest.fixture
-def github():
-    ROOT_REPO = DummyRepository(
-        "SFDO-Tooling",
-        "RootRepo",
-        {
-            "cumulusci.yml": DummyContents(
-                b"""
-    project:
-        name: RootRepo
-        package:
-            name: RootRepo
-            namespace: bar
-        git:
-            repo_url: https://github.com/SFDO-Tooling/CumulusCI-Test
-        dependencies:
-            - github: https://github.com/SFDO-Tooling/DependencyRepo
-    """
-            ),
-            "unpackaged/pre": {"first": {}, "second": {}},
-            "src": {"src": ""},
-            "unpackaged/post": {"first": {}},
-        },
-        [
-            DummyRelease("release/2.0", "2.0"),
-        ],
-    )
-
-    DEPENDENCY_REPO = DummyRepository(
-        "SFDO-Tooling",
-        "DependencyRepo",
-        {
-            "cumulusci.yml": DummyContents(
-                b"""
-    project:
-        name: DependencyRepo
-        package:
-            name: DependencyRepo
-            namespace: foo
-        git:
-            repo_url: https://github.com/SFDO-Tooling/CumulusCI-Test
-    """
-            ),
-            "unpackaged/pre": {"top": {}},
-            "src": {"src": ""},
-            "unpackaged/post": {"top": {}},
-        },
-        [
-            DummyRelease("release/1.1", "1.1"),
-        ],
-    )
-
-    # This repo contains an unparseable transitive dependency.
-    ROOT_BAD_DEP_REPO = DummyRepository(
-        "SFDO-Tooling",
-        "RootRepoBadDep",
-        {
-            "cumulusci.yml": DummyContents(
-                b"""
-    project:
-        name: RootRepo
-        package:
-            name: RootRepo
-            namespace: bar
-        git:
-            repo_url: https://github.com/SFDO-Tooling/CumulusCI-Test
-        dependencies:
-            - bogus: foo
-    """
-            ),
-        },
-    )
-
-    # This repo contains both beta and managed releases.
-    RELEASES_REPO = DummyRepository(
-        "SFDO-Tooling",
-        "CumulusCI-Test-Dep",
-        {
-            "cumulusci.yml": DummyContents(
-                b"""
-    project:
-        name: CumulusCI-Test-Dep
-        package:
-            name: CumulusCI-Test-Dep
-            namespace: ccitestdep
-        git:
-            repo_url: https://github.com/SFDO-Tooling/CumulusCI-Test-Dep
-    """
-            ),
-            "unpackaged/pre": {},
-            "src": {},
-            "unpackaged/post": {},
-        },
-        [
-            DummyRelease("beta/2.1_Beta_1", "2.1 Beta 1"),
-            DummyRelease("release/2.0", "2.0"),
-            DummyRelease("release/1.0", "1.0"),
-        ],
-    )
-
-    # This repo contains releases, but no `unmanaged/pre`
-    NO_UNMANAGED_PRE_REPO = DummyRepository(
-        "SFDO-Tooling",
-        "NoUnmanagedPreRepo",
-        {
-            "cumulusci.yml": DummyContents(
-                b"""
-    project:
-        name: NoUnmanagedPreRepo
-        package:
-            name: NoUnmanagedPreRepo
-            namespace: foo
-        git:
-            repo_url: https://github.com/SFDO-Tooling/NoUnmanagedPreRepo
-    """
-            ),
-            "src": {},
-            "unpackaged/post": {},
-        },
-        [
-            DummyRelease("beta/2.1_Beta_1", "2.1 Beta 1"),
-            DummyRelease("release/2.0", "2.0"),
-            DummyRelease("release/1.0", "1.0"),
-        ],
-    )
-
-    CUMULUSCI_REPO = DummyRepository(
-        "SFDO-Tooling",
-        "CumulusCI",
-        {},
-        [
-            DummyRelease("release/1.1", "1.1"),
-            DummyRelease("beta-wrongprefix", "wrong"),
-            DummyRelease("release/1.0", "1.0"),
-            DummyRelease("beta/1.0-Beta_2", "1.0 (Beta 2)"),
-            DummyRelease("beta/1.0-Beta_1", "1.0 (Beta 1)"),
-        ],
-    )
-
-    # This repo contains a release, but no namespace
-    UNMANAGED_REPO = DummyRepository(
-        "SFDO-Tooling",
-        "UnmanagedRepo",
-        {
-            "cumulusci.yml": DummyContents(
-                b"""
-    project:
-        name: CumulusCI-Test
-        package:
-            name: CumulusCI-Test
-    """
-            ),
-            "unpackaged/pre": {"pre": {}, "skip": {}},
-            "src": {"src": ""},
-            "unpackaged/post": {"post": {}, "skip": {}},
-        },
-        [
-            DummyRelease("release/1.0", "1.0"),
-        ],
-    )
-
-    TWO_GP_REPO = DummyRepository(
-        "SFDO-Tooling",
-        "TwoGPRepo",
-        {
-            "cumulusci.yml": DummyContents(
-                b"""
-    project:
-        name: CumulusCI-2GP-Test
-        package:
-            name: CumulusCI-2GP-Test
-        git:
-            2gp_context: "Nonstandard Package Status"
-    """
-            ),
-            "unpackaged/pre": {"pre": {}, "skip": {}},
-            "src": {"src": ""},
-            "unpackaged/post": {"post": {}, "skip": {}},
-        },
-        releases=[],
-        commits={
-            "main_sha": mock.Mock(sha="main_sha"),
-            "feature/232_sha": mock.Mock(
-                sha="feature/232_sha",
-                parents=[{"sha": "parent_sha"}],
-                status=mock.Mock(return_value=mock.Mock(statuses=[])),
-            ),
-            "parent_sha": mock.Mock(
-                sha="parent_sha",
-                parents=[],
-                status=mock.Mock(
-                    return_value=mock.Mock(
-                        statuses=[
-                            mock.Mock(
-                                state="success",
-                                context="Nonstandard Package Status",
-                                description="version_id: 04t000000000000",
-                            )
-                        ]
-                    )
-                ),
-            ),
-            "feature/232__test_sha": mock.Mock(
-                sha="feature/232__test_sha",
-                parents=[],
-                status=mock.Mock(
-                    return_value=mock.Mock(
-                        statuses=[
-                            mock.Mock(
-                                state="success",
-                                context="Nonstandard Package Status",
-                                description="version_id: 04t000000000001",
-                            )
-                        ]
-                    )
-                ),
-            ),
-        },
-    )
-
-    TWO_GP_MISSING_REPO = DummyRepository(
-        "SFDO-Tooling",
-        "TwoGPMissingRepo",
-        {
-            "cumulusci.yml": DummyContents(
-                b"""
-    project:
-        name: CumulusCI-2GP-Test
-        package:
-            name: CumulusCI-2GP-Test
-        git:
-            2gp_context: "Nonstandard Package Status"
-    """
-            ),
-            "unpackaged/pre": {"pre": {}, "skip": {}},
-            "src": {"src": ""},
-            "unpackaged/post": {"post": {}, "skip": {}},
-        },
-        releases=[],
-        commits={
-            "main_sha": mock.Mock(sha="main_sha"),
-            "feature/232_sha": mock.Mock(
-                sha="feature/232_sha",
-                parents=[{"sha": "parent_sha"}],
-                status=mock.Mock(return_value=mock.Mock(statuses=[])),
-            ),
-            "parent_sha": mock.Mock(
-                sha="parent_sha",
-                parents=[],
-                status=mock.Mock(return_value=mock.Mock(statuses=[])),
-            ),
-        },
-    )
-
-    def branch(which_repo, which_branch):
-        branch = mock.Mock()
-        branch.commit = which_repo.commit(f"{which_branch}_sha")
-        return branch
-
-    TWO_GP_REPO.branch = mock.Mock(wraps=functools.partial(branch, TWO_GP_REPO))
-    TWO_GP_MISSING_REPO.branch = mock.Mock(
-        wraps=functools.partial(branch, TWO_GP_MISSING_REPO)
-    )
-
-    return DummyGithub(
-        {
-            "UnmanagedRepo": UNMANAGED_REPO,
-            "CumulusCI": CUMULUSCI_REPO,
-            "NoUnmanagedPreRepo": NO_UNMANAGED_PRE_REPO,
-            "RootRepo": ROOT_REPO,
-            "RootRepoBadDep": ROOT_BAD_DEP_REPO,
-            "DependencyRepo": DEPENDENCY_REPO,
-            "ReleasesRepo": RELEASES_REPO,
-            "TwoGPRepo": TWO_GP_REPO,
-            "TwoGPMissingRepo": TWO_GP_MISSING_REPO,
-        }
-    )
-
-
-@pytest.fixture
-def project_config(github):
-    pc = mock.Mock()
-
-    def get_repo_from_url(url):
-        return github.repository(*split_repo_url(url))
-
-    pc.get_repo_from_url = get_repo_from_url
-
-    return pc
 
 
 class ConcreteDynamicDependency(DynamicDependency):
@@ -390,7 +78,7 @@ class MockBadResolver(Resolver):
 
 
 class TestDynamicDependency:
-    @mock.patch("cumulusci.core.dependencies.dependencies.get_resolver")
+    @mock.patch("cumulusci.core.dependencies.resolvers.get_resolver")
     def test_dynamic_dependency(self, get_resolver):
         d = ConcreteDynamicDependency()
         resolvers = [
@@ -415,7 +103,7 @@ class TestDynamicDependency:
         )
         assert d.ref == "aaaaaaaaaaaaaaaa"
 
-    @mock.patch("cumulusci.core.dependencies.dependencies.get_resolver")
+    @mock.patch("cumulusci.core.dependencies.resolvers.get_resolver")
     def test_dynamic_dependency__twice(self, get_resolver):
         d = ConcreteDynamicDependency()
         resolvers = [
@@ -448,7 +136,7 @@ class TestDynamicDependency:
         assert d.is_resolved
         resolvers[0].resolve.assert_called_once()
 
-    @mock.patch("cumulusci.core.dependencies.dependencies.get_resolver")
+    @mock.patch("cumulusci.core.dependencies.resolvers.get_resolver")
     def test_dynamic_dependency_resolution_fails(self, get_resolver):
         d = ConcreteDynamicDependency()
         resolvers = [MockBadResolver()]
@@ -457,7 +145,7 @@ class TestDynamicDependency:
         with pytest.raises(DependencyResolutionError):
             d.resolve(mock.Mock(), [DependencyResolutionStrategy.UNMANAGED_HEAD])
 
-    @mock.patch("cumulusci.core.dependencies.dependencies.get_resolver")
+    @mock.patch("cumulusci.core.dependencies.resolvers.get_resolver")
     def test_dynamic_dependency_resolution_no_results(self, get_resolver):
         d = ConcreteDynamicDependency()
         resolvers = [MockResolver("", None)]
@@ -1063,505 +751,3 @@ class TestParseDependency:
             }
         )
         assert isinstance(u, UnmanagedZipURLDependency)
-
-
-class TestGitHubTagResolver:
-    def test_github_tag_resolver(self, project_config):
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/ReleasesRepo",
-            tag="release/1.0",  # Not the most recent release
-        )
-        resolver = GitHubTagResolver()
-
-        assert resolver.can_resolve(dep, project_config)
-        assert resolver.resolve(dep, project_config) == (
-            "tag_sha",
-            PackageNamespaceVersionDependency(
-                namespace="ccitestdep", version="1.0", package_name="CumulusCI-Test-Dep"
-            ),
-        )
-
-    def test_github_tag_resolver__unmanaged(self, project_config):
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/ReleasesRepo",
-            tag="release/2.0",
-            unmanaged=True,
-        )
-        resolver = GitHubTagResolver()
-
-        assert resolver.resolve(dep, project_config) == (
-            "tag_sha",
-            None,
-        )
-
-    def test_exception_no_managed_release(self, project_config):
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/UnmanagedRepo",  # This repo has no namespace
-            tag="release/1.0",
-            unmanaged=False,
-        )
-        resolver = GitHubTagResolver()
-
-        with pytest.raises(DependencyResolutionError) as e:
-            resolver.resolve(dep, project_config)
-        assert "does not identify a managed release" in str(e.value)
-
-    def test_can_resolve_negative(self, project_config):
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/ReleasesRepo"
-        )
-        resolver = GitHubTagResolver()
-
-        assert not resolver.can_resolve(dep, project_config)
-
-    def test_exception_no_tag_found(self, project_config):
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/ReleasesRepo",
-            tag="release/3.0",
-        )
-        resolver = GitHubTagResolver()
-
-        with pytest.raises(DependencyResolutionError) as e:
-            resolver.resolve(dep, project_config)
-        assert "No release found for tag" in str(e.value)
-
-    def test_str(self):
-        assert str(GitHubTagResolver()) == GitHubTagResolver.name
-
-
-class TestGitHubReleaseTagResolver:
-    def test_github_release_tag_resolver(self, project_config):
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/ReleasesRepo"
-        )
-        resolver = GitHubReleaseTagResolver()
-
-        assert resolver.can_resolve(dep, project_config)
-        assert resolver.resolve(dep, project_config) == (
-            "tag_sha",
-            PackageNamespaceVersionDependency(
-                namespace="ccitestdep", version="2.0", package_name="CumulusCI-Test-Dep"
-            ),
-        )
-
-    def test_beta_release_tag(self, project_config):
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/ReleasesRepo"
-        )
-        resolver = GitHubBetaReleaseTagResolver()
-
-        assert resolver.can_resolve(dep, project_config)
-        assert resolver.resolve(dep, project_config) == (
-            "tag_sha",
-            PackageNamespaceVersionDependency(
-                namespace="ccitestdep",
-                version="2.1 Beta 1",
-                package_name="CumulusCI-Test-Dep",
-            ),
-        )
-
-    def test_not_found(self, project_config):
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/TwoGPRepo"  # This mock repo contains no releases
-        )
-        resolver = GitHubReleaseTagResolver()
-
-        assert resolver.can_resolve(dep, project_config)
-        assert resolver.resolve(dep, project_config) == (None, None)
-
-
-class TestGitHubUnmanagedHeadResolver:
-    def test_unmanaged_head_resolver(self, project_config):
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/ReleasesRepo"
-        )
-        resolver = GitHubUnmanagedHeadResolver()
-
-        assert resolver.can_resolve(dep, project_config)
-
-        assert resolver.resolve(dep, project_config) == ("commit_sha", None)
-
-
-class ConcreteGitHubReleaseBranchResolver(GitHubReleaseBranchResolver):
-    def resolve(
-        self, dep: GitHubDynamicDependency, context: BaseProjectConfig
-    ) -> Tuple[Optional[str], Optional[StaticDependency]]:
-        return (None, None)
-
-
-class TestGitHubReleaseBranchResolver:
-    def test_is_valid_repo_context(self):
-        pc = BaseProjectConfig(UniversalConfig())
-
-        pc.repo_info["branch"] = "feature/232__test"
-        pc.project__git["prefix_feature"] = "feature/"
-
-        assert ConcreteGitHubReleaseBranchResolver().is_valid_repo_context(pc)
-
-        pc.repo_info["branch"] = "main"
-        assert not ConcreteGitHubReleaseBranchResolver().is_valid_repo_context(pc)
-
-    def test_can_resolve(self):
-        pc = BaseProjectConfig(UniversalConfig())
-
-        pc.repo_info["branch"] = "feature/232__test"
-        pc.project__git["prefix_feature"] = "feature/"
-
-        gh = ConcreteGitHubReleaseBranchResolver()
-
-        assert gh.can_resolve(
-            GitHubDynamicDependency(github="https://github.com/SFDO-Tooling/Test"),
-            pc,
-        )
-
-        assert not gh.can_resolve(ConcreteDynamicDependency(), pc)
-
-    def test_get_release_id(self):
-        pc = BaseProjectConfig(UniversalConfig())
-        pc.repo_info["branch"] = "feature/232__test"
-        pc.project__git["prefix_feature"] = "feature/"
-
-        assert ConcreteGitHubReleaseBranchResolver().get_release_id(pc) == 232
-
-    def test_get_release_id__not_release_branch(self):
-        pc = BaseProjectConfig(UniversalConfig())
-        with mock.patch.object(
-            BaseProjectConfig, "repo_branch", new_callable=mock.PropertyMock
-        ) as repo_branch:
-            repo_branch.return_value = None
-
-            with pytest.raises(DependencyResolutionError) as e:
-                ConcreteGitHubReleaseBranchResolver().get_release_id(pc)
-
-            assert "Cannot get current branch" in str(e)
-
-    def test_get_release_id__no_git_data(self):
-        pc = BaseProjectConfig(UniversalConfig())
-        with mock.patch.object(
-            BaseProjectConfig, "repo_branch", new_callable=mock.PropertyMock
-        ) as repo_branch:
-            repo_branch.return_value = "feature/test"
-            pc.project__git["prefix_feature"] = "feature/"
-
-            with pytest.raises(DependencyResolutionError) as e:
-                ConcreteGitHubReleaseBranchResolver().get_release_id(pc)
-
-            assert "Cannot get current release identifier" in str(e)
-
-    def test_locate_commit_status_package_id__not_found_with_parent(self, github):
-        repo = github.repository("SFDO-Tooling", "TwoGPMissingRepo")
-        assert ConcreteGitHubReleaseBranchResolver().locate_commit_status_package_id(
-            repo, repo.branch("feature/232"), "Build Feature Test Package"
-        ) == (None, None)
-
-
-class TestGitHubReleaseBranchCommitStatusResolver:
-    def test_2gp_release_branch_resolver(self, project_config):
-        project_config.repo_branch = "feature/232__test"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/TwoGPRepo"
-        )
-
-        assert resolver.can_resolve(dep, project_config)
-
-        assert resolver.resolve(dep, project_config) == (
-            "parent_sha",
-            PackageVersionIdDependency(
-                version_id="04t000000000000", package_name="CumulusCI-2GP-Test"
-            ),
-        )
-
-    def test_commit_status_not_found(self, project_config):
-        project_config.repo_branch = "feature/232__test"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/TwoGPMissingRepo"
-        )
-
-        assert resolver.can_resolve(dep, project_config)
-        assert resolver.resolve(dep, project_config) == (None, None)
-
-    def test_repo_not_found(self, project_config):
-        project_config.repo_branch = "feature/232__test"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/NonexistentRepo"
-        )
-
-        project_config.get_repo_from_url = mock.Mock(return_value=None)
-
-        with pytest.raises(DependencyResolutionError) as exc:
-            resolver.resolve(dep, project_config)
-
-        assert "Unable to access GitHub" in str(exc)
-
-    @mock.patch("cumulusci.core.dependencies.dependencies.find_repo_feature_prefix")
-    def test_unable_locate_feature_prefix(
-        self, find_repo_feature_prefix_mock, project_config
-    ):
-        find_repo_feature_prefix_mock.side_effect = NotFoundError
-        project_config.repo_branch = "feature/232__test"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/TwoGPRepo"
-        )
-        assert resolver.resolve(dep, project_config) == (None, None)
-
-    def test_branch_not_found(self, project_config):
-        project_config.repo_branch = "feature/290__test"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/TwoGPRepo"
-        )
-
-        assert resolver.resolve(dep, project_config) == (None, None)
-
-
-class TestGitHubReleaseBranchExactMatchCommitStatusResolver:
-    def test_exact_branch_resolver(self, project_config):
-        project_config.repo_branch = "feature/232__test"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchExactMatchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/TwoGPRepo"
-        )
-
-        assert resolver.can_resolve(dep, project_config)
-
-        assert resolver.resolve(dep, project_config) == (
-            "feature/232__test_sha",
-            PackageVersionIdDependency(
-                version_id="04t000000000001", package_name="CumulusCI-2GP-Test"
-            ),
-        )
-
-    def test_repo_not_found(self, project_config):
-        project_config.repo_branch = "feature/232__test"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchExactMatchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/NonexistentRepo"
-        )
-
-        project_config.get_repo_from_url = mock.Mock(return_value=None)
-
-        with pytest.raises(DependencyResolutionError) as exc:
-            resolver.resolve(dep, project_config)
-
-        assert "Unable to access GitHub" in str(exc)
-
-    @mock.patch("cumulusci.core.dependencies.dependencies.find_repo_feature_prefix")
-    def test_unable_locate_feature_prefix(
-        self, find_repo_feature_prefix_mock, project_config
-    ):
-        find_repo_feature_prefix_mock.side_effect = NotFoundError
-        project_config.repo_branch = "feature/232__test"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchExactMatchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/TwoGPRepo"
-        )
-        assert resolver.resolve(dep, project_config) == (None, None)
-
-    def test_branch_not_found(self, project_config):
-        project_config.repo_branch = "feature/290__test"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchExactMatchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/TwoGPRepo"
-        )
-
-        assert resolver.resolve(dep, project_config) == (None, None)
-
-    def test_commit_status_not_found(self, project_config):
-        project_config.repo_branch = "feature/232"
-        project_config.project__git__prefix_feature = "feature/"
-
-        resolver = GitHubReleaseBranchExactMatchCommitStatusResolver()
-        dep = GitHubDynamicDependency(
-            github="https://github.com/SFDO-Tooling/TwoGPMissingRepo"
-        )
-
-        assert resolver.can_resolve(dep, project_config)
-        assert resolver.resolve(dep, project_config) == (None, None)
-
-
-class TestResolverAccess:
-    def test_get_resolver(self):
-        assert isinstance(
-            get_resolver(
-                DependencyResolutionStrategy.STATIC_TAG_REFERENCE,
-                GitHubDynamicDependency(github="https://github.com/SFDO-Tooling/Test"),
-            ),
-            GitHubTagResolver,
-        )
-
-    def test_get_resolver_stack__indirect(self):
-        pc = BaseProjectConfig(UniversalConfig())
-
-        strategy = get_resolver_stack(pc, "production")
-        assert DependencyResolutionStrategy.RELEASE_TAG in strategy
-        assert DependencyResolutionStrategy.BETA_RELEASE_TAG not in strategy
-
-    def test_get_resolver_stack__customized_indirect(self):
-        pc = BaseProjectConfig(UniversalConfig())
-
-        pc.project__dependency_resolutions["preproduction"] = "include_beta"
-        strategy = get_resolver_stack(pc, "preproduction")
-        assert DependencyResolutionStrategy.BETA_RELEASE_TAG in strategy
-
-    def test_get_resolver_stack__direct(self):
-        pc = BaseProjectConfig(UniversalConfig())
-
-        strategy = get_resolver_stack(pc, "commit_status")
-        assert DependencyResolutionStrategy.COMMIT_STATUS_RELEASE_BRANCH in strategy
-
-    def test_get_resolver_stack__fail(self):
-        pc = BaseProjectConfig(UniversalConfig())
-
-        with pytest.raises(CumulusCIException) as e:
-            get_resolver_stack(pc, "bogus")
-
-        assert "not found" in str(e)
-
-
-class TestStaticDependencyResolution:
-    def test_get_static_dependencies(self, project_config):
-        gh = GitHubDynamicDependency(github="https://github.com/SFDO-Tooling/RootRepo")
-
-        assert get_static_dependencies(
-            [gh], [DependencyResolutionStrategy.RELEASE_TAG], project_config
-        ) == [
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/DependencyRepo",
-                subfolder="unpackaged/pre/top",
-                unmanaged=True,
-                ref="tag_sha",
-            ),
-            PackageNamespaceVersionDependency(
-                namespace="foo", version="1.1", package_name="DependencyRepo"
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/DependencyRepo",
-                subfolder="unpackaged/post/top",
-                unmanaged=False,
-                ref="tag_sha",
-                namespace_inject="foo",
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/RootRepo",
-                subfolder="unpackaged/pre/first",
-                unmanaged=True,
-                ref="tag_sha",
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/RootRepo",
-                subfolder="unpackaged/pre/second",
-                unmanaged=True,
-                ref="tag_sha",
-            ),
-            PackageNamespaceVersionDependency(
-                namespace="bar", version="2.0", package_name="RootRepo"
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/RootRepo",
-                subfolder="unpackaged/post/first",
-                unmanaged=False,
-                ref="tag_sha",
-                namespace_inject="bar",
-            ),
-        ]
-
-    def test_get_static_dependencies__ignore_namespace(self, project_config):
-        gh = GitHubDynamicDependency(github="https://github.com/SFDO-Tooling/RootRepo")
-
-        assert get_static_dependencies(
-            [gh],
-            [DependencyResolutionStrategy.RELEASE_TAG],
-            project_config,
-            ignore_deps=[{"namespace": "foo"}],
-        ) == [
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/DependencyRepo",
-                subfolder="unpackaged/pre/top",
-                unmanaged=True,
-                ref="tag_sha",
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/DependencyRepo",
-                subfolder="unpackaged/post/top",
-                unmanaged=False,
-                ref="tag_sha",
-                namespace_inject="foo",
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/RootRepo",
-                subfolder="unpackaged/pre/first",
-                unmanaged=True,
-                ref="tag_sha",
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/RootRepo",
-                subfolder="unpackaged/pre/second",
-                unmanaged=True,
-                ref="tag_sha",
-            ),
-            PackageNamespaceVersionDependency(
-                namespace="bar", version="2.0", package_name="RootRepo"
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/RootRepo",
-                subfolder="unpackaged/post/first",
-                unmanaged=False,
-                ref="tag_sha",
-                namespace_inject="bar",
-            ),
-        ]
-
-    def test_get_static_dependencies__ignore_github(self, project_config):
-        gh = GitHubDynamicDependency(github="https://github.com/SFDO-Tooling/RootRepo")
-
-        assert get_static_dependencies(
-            [gh],
-            [DependencyResolutionStrategy.RELEASE_TAG],
-            project_config,
-            ignore_deps=[{"github": "https://github.com/SFDO-Tooling/DependencyRepo"}],
-        ) == [
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/RootRepo",
-                subfolder="unpackaged/pre/first",
-                unmanaged=True,
-                ref="tag_sha",
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/RootRepo",
-                subfolder="unpackaged/pre/second",
-                unmanaged=True,
-                ref="tag_sha",
-            ),
-            PackageNamespaceVersionDependency(
-                namespace="bar", version="2.0", package_name="RootRepo"
-            ),
-            UnmanagedGitHubRefDependency(
-                github="https://github.com/SFDO-Tooling/RootRepo",
-                subfolder="unpackaged/post/first",
-                unmanaged=False,
-                ref="tag_sha",
-                namespace_inject="bar",
-            ),
-        ]
